match ee code

+++ api/http/models/kubernetes/dashboard.go
@@ -2,12 +2,12 @@ package kubernetes
 
 type (
 	K8sDashboard struct {
-		NamespacesCount   int `json:"namespacesCount"`
-		ApplicationsCount int `json:"applicationsCount"`
-		ServicesCount     int `json:"servicesCount"`
-		IngressesCount    int `json:"ingressesCount"`
-		ConfigMapsCount   int `json:"configMapsCount"`
-		SecretsCount      int `json:"secretsCount"`
-		VolumesCount      int `json:"volumesCount"`
+		NamespacesCount   int64 `json:"namespacesCount"`
+		ApplicationsCount int64 `json:"applicationsCount"`
+		ServicesCount     int64 `json:"servicesCount"`
+		IngressesCount    int64 `json:"ingressesCount"`
+		ConfigMapsCount   int64 `json:"configMapsCount"`
+		SecretsCount      int64 `json:"secretsCount"`
+		VolumesCount      int64 `json:"volumesCount"`
 	}
 )
